+++ device-bootloaders/CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - [SW-226] Added initial support for GET and SET FRU commands
 - [SW-20541]: use a dedicated BL2_CODE_REGION_SIZE per BL2 build target
 - [SW-20543]: implement SP FRU commands
+- [SW-19030] Extended flash config to have minion, sram and noc voltage values
 ### Changed
 - Optimize NOC remap algorithm
 - Make FRU data raw binary